LPD-61591 Move tests to new file
